requirements.txt

langchain>=0.2.0
langchain-community>=0.2.16
langchain-google-genai>=1.0.7
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0


faiss-cpu>=1.8.0
chromadb>=0.5.0


PyPDF2>=3.0.0
pypdf>=4.2.0


sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.1.0

google-generativeai>=0.7.0

Python-dotenv>=1.0.0


rich>=13.7.0
tabulate>=0.9.0
